# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: POS Kafe Multi-Cabang API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /reservations:
    get:
      operationId: getReservations
      tags: [reservations]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Reservation"
    post:
      operationId: createReservation
      tags: [reservations]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationInput"
      responses:
        "201":
          description: Reservation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
  /reservations/{id}:
    get:
      operationId: getReservation
      tags: [reservations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Reservation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    patch:
      operationId: updateReservation
      tags: [reservations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReservationUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Reservation"
    delete:
      operationId: deleteReservation
      tags: [reservations]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ORDERS
...
  /reports/sales-summary:
    get:
      operationId: getSalesSummary
      tags: [reports]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: groupBy
          in: query
          schema:
            type: string
            enum: [day, week, month]
      responses:
        "200":
          description: Sales summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SalesSummary"
  /reports/top-items:
    get:
      operationId: getTopMenuItems
      tags: [reports]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Top menu items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TopMenuItem"
  /reports/branch-comparison:
    get:
      operationId: getBranchComparison
      tags: [reports]
      parameters:
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Branch comparison
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BranchStat"
  /reports/payment-methods:
    get:
      operationId: getPaymentMethodStats
      tags: [reports]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: startDate
          in: query
          required: true
          schema:
            type: string
        - name: endDate
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment method breakdown
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PaymentMethodStat"

  # DASHBOARD
  /dashboard/overview:
    get:
      operationId: getDashboardOverview
      tags: [dashboard]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Dashboard overview
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardOverview"
  /dashboard/peak-hours:
    get:
      operationId: getPeakHours
      tags: [dashboard]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Peak hours data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PeakHourStat"
  /dashboard/low-stock:
    get:
      operationId: getLowStockAlerts
      tags: [dashboard]
      parameters:
        - name: branchId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Low stock alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/LowStockAlert"

  # UPLOAD
...
components:
  schemas:
...
    Reservation:
      type: object
      required: [id, customerName, customerPhone, date, time, guestCount, status, branchId, createdAt]
      properties:
        id:
          type: integer
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        guestCount:
          type: integer
        tableId:
          type: ["integer", "null"]
        tableNumber:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, confirmed, seated, completed, cancelled, waitlist]
        notes:
          type: ["string", "null"]
        depositAmount:
          type: ["number", "null"]
        branchId:
          type: integer
        createdAt:
          type: string
    ReservationInput:
      type: object
      required: [customerName, customerPhone, date, time, guestCount, branchId]
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        guestCount:
          type: integer
        tableId:
          type: ["integer", "null"]
        notes:
          type: ["string", "null"]
        depositAmount:
          type: ["number", "null"]
        branchId:
          type: integer
    ReservationUpdate:
      type: object
      properties:
        customerName:
          type: string
        customerPhone:
          type: string
        customerEmail:
          type: ["string", "null"]
        date:
          type: string
        time:
          type: string
        guestCount:
          type: integer
        tableId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [pending, confirmed, seated, completed, cancelled, waitlist]
        notes:
          type: ["string", "null"]
        depositAmount:
          type: ["number", "null"]

    # ORDER
...
    SalesSummary:
      type: object
      required: [totalRevenue, totalOrders, averageOrderValue, periods]
      properties:
        totalRevenue:
          type: number
        totalOrders:
          type: integer
        averageOrderValue:
          type: number
        periods:
          type: array
          items:
            $ref: "#/components/schemas/SalesPeriod"
    SalesPeriod:
      type: object
      required: [date, revenue, orders]
      properties:
        date:
          type: string
        revenue:
          type: number
        orders:
          type: integer
    TopMenuItem:
      type: object
      required: [menuItemId, menuItemName, quantitySold, totalRevenue]
      properties:
        menuItemId:
          type: integer
        menuItemName:
          type: string
        imageUrl:
          type: ["string", "null"]
        quantitySold:
          type: integer
        totalRevenue:
          type: number
    BranchStat:
      type: object
      required: [branchId, branchName, revenue, orders, averageOrderValue]
      properties:
        branchId:
          type: integer
        branchName:
          type: string
        revenue:
          type: number
        orders:
          type: integer
        averageOrderValue:
          type: number
    PaymentMethodStat:
      type: object
      required: [method, count, amount, percentage]
      properties:
        method:
          type: string
        count:
          type: integer
        amount:
          type: number
        percentage:
          type: number

    # DASHBOARD
    DashboardOverview:
      type: object
      required: [todayRevenue, todayOrders, activeOrders, activeTables, totalTables, pendingReservations, lowStockCount]
      properties:
        todayRevenue:
          type: number
        todayOrders:
          type: integer
        activeOrders:
          type: integer
        activeTables:
          type: integer
        totalTables:
          type: integer
        pendingReservations:
          type: integer
        lowStockCount:
          type: integer
        revenueChange:
          type: number
        ordersChange:
          type: number
    PeakHourStat:
      type: object
      required: [hour, orders, revenue]
      properties:
        hour:
          type: integer
        orders:
          type: integer
        revenue:
          type: number
    LowStockAlert:
      type: object
      required: [ingredientId, ingredientName, currentStock, minStock, unit]
      properties:
        ingredientId:
          type: integer
        ingredientName:
          type: string
        currentStock:
          type: number
        minStock:
          type: number
        unit:
          type: string

    # UPLOAD
...